Update landing plugin to support restoration.
